# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Trackable Ingress API
  description: |
    Personal shopping assistant API for post-purchase order management.

    This API provides conversational AI capabilities for managing online shopping orders,
    tracking shipments, understanding merchant policies, and monitoring return/exchange deadlines.

    **Authentication:** This is an internal Cloud Run service requiring GCP IAM authentication.
    Include an identity token in the Authorization header: `Bearer <token>`

    **Model:** Powered by Gemini 2.5 Flash via Google ADK
  version: 0.1.0
  contact:
    name: Trackable Team
  license:
    name: Proprietary

security:
  - BearerAuth: []

tags:
  - name: system
    description: System health and information endpoints
  - name: chat
    description: OpenAI-compatible chat completions API
  - name: ingest
    description: Manual email and image submission endpoints

paths:
  /:
    get:
      tags:
        - system
      summary: Get service information
      description: Returns basic information about the API service
      operationId: getServiceInfo
      responses:
        "200":
          description: Service information
          content:
            application/json:
              schema:
                type: object
                properties:
                  service:
                    type: string
                    example: Trackable Ingress API
                  version:
                    type: string
                    example: 0.1.0
                  status:
                    type: string
                    example: operational
                  description:
                    type: string
                    example: Personal shopping assistant for post-purchase management

  /health:
    get:
      tags:
        - system
      summary: Health check
      description: Check service health status (used by Cloud Run monitoring)
      operationId: healthCheck
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  service:
                    type: string
                    example: trackable-ingress
                  environment:
                    type: string
                    example: gen-lang-client-0659747538

  /api/v1/chat/completions:
    post:
      tags:
        - chat
      summary: Create chat completion (OpenAI-compatible)
      description: |
        Create a chat completion following the OpenAI API format.

        This endpoint provides conversational AI capabilities using Google ADK and Gemini 2.5 Flash,
        wrapped in an OpenAI-compatible interface for easy integration with existing tools.

        **Streaming:**
        Set `stream: true` to receive Server-Sent Events (SSE) with incremental responses.

        **Session Management:**
        Sessions are managed per-user via the `user` field. Include the same `user` value
        in subsequent requests to maintain conversation context.
      operationId: createChatCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
            examples:
              basic:
                summary: Basic chat request
                value:
                  model: "gemini-2.5-flash"
                  messages:
                    - role: "user"
                      content: "Hello! What is Trackable?"
                  user: "user_12345"
              withSystem:
                summary: With system message
                value:
                  model: "gemini-2.5-flash"
                  messages:
                    - role: "system"
                      content: "You are a helpful shopping assistant."
                    - role: "user"
                      content: "What can you help me with?"
                  user: "user_12345"
              streaming:
                summary: Streaming request
                value:
                  model: "gemini-2.5-flash"
                  messages:
                    - role: "user"
                      content: "Tell me about order tracking"
                  stream: true
                  user: "user_12345"
      responses:
        "200":
          description: Successful completion (non-streaming)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
              example:
                id: "chatcmpl-abc123def456"
                object: "chat.completion"
                created: 1706300000
                model: "gemini-2.5-flash"
                choices:
                  - index: 0
                    message:
                      role: "assistant"
                      content: "Hello! I'm Trackable, your personal shopping assistant."
                    finish_reason: "stop"
                usage:
                  prompt_tokens: 10
                  completion_tokens: 15
                  total_tokens: 25
            text/event-stream:
              schema:
                type: string
                description: |
                  SSE stream of chat completion chunks (when stream=true).

                  Each chunk follows the format:
                  ```
                  data: {"id":"chatcmpl-...","choices":[{"delta":{"content":"..."}}]}
                  ```

                  Stream ends with:
                  ```
                  data: [DONE]
                  ```
              examples:
                streamResponse:
                  summary: Streaming response
                  value: |
                    data: {"id":"chatcmpl-abc123","object":"chat.completion.chunk","created":1706300000,"model":"gemini-2.5-flash","choices":[{"index":0,"delta":{"role":"assistant"},"finish_reason":null}]}

                    data: {"id":"chatcmpl-abc123","object":"chat.completion.chunk","created":1706300000,"model":"gemini-2.5-flash","choices":[{"index":0,"delta":{"content":"Hello!"},"finish_reason":null}]}

                    data: {"id":"chatcmpl-abc123","object":"chat.completion.chunk","created":1706300000,"model":"gemini-2.5-flash","choices":[{"index":0,"delta":{"content":" I'm Trackable."},"finish_reason":null}]}

                    data: {"id":"chatcmpl-abc123","object":"chat.completion.chunk","created":1706300000,"model":"gemini-2.5-flash","choices":[{"index":0,"delta":{},"finish_reason":"stop"}]}

                    data: [DONE]
          headers:
            Cache-Control:
              schema:
                type: string
                example: no-cache
            Connection:
              schema:
                type: string
                example: keep-alive
            X-Accel-Buffering:
              schema:
                type: string
                example: "no"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: "messages: field required"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: "Chat completion failed: Internal error"

  /api/v1/chat/sessions:
    delete:
      tags:
        - chat
      summary: Clear chat session
      description: |
        Clear the chat session for a user, resetting conversation history.

        After clearing, the next chat request from the user will automatically
        create a fresh session with no prior context.
      operationId: clearChatSession
      parameters:
        - name: user
          in: query
          description: User identifier (defaults to "default_user")
          required: false
          schema:
            type: string
            default: "default_user"
            example: "user_12345"
      responses:
        "200":
          description: Session cleared or not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClearSessionResponse"
              examples:
                cleared:
                  summary: Session cleared
                  value:
                    message: "Session cleared"
                    user: "user_12345"
                notFound:
                  summary: No session found
                  value:
                    message: "No session found"
                    user: "user_12345"

  /api/v1/ingest/email:
    post:
      tags:
        - ingest
      summary: Submit email for order extraction
      description: |
        Submit an email for order extraction. Creates a job that is processed
        asynchronously by the Worker service via Cloud Tasks.

        The email content will be parsed by the input processor agent to extract
        order information including merchant, items, prices, and tracking details.
      operationId: ingestEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestEmailRequest"
            examples:
              basic:
                summary: Basic email submission
                value:
                  email_content: "From: orders@amazon.com\nSubject: Your order has shipped\n\nOrder #123-456 has shipped..."
              withMetadata:
                summary: With metadata
                value:
                  email_content: "Order confirmation from Nike..."
                  email_subject: "Your Nike order confirmation"
                  email_from: "orders@nike.com"
      responses:
        "200":
          description: Email submitted for processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
              example:
                job_id: "550e8400-e29b-41d4-a716-446655440000"
                source_id: "660e8400-e29b-41d4-a716-446655440001"
                status: "queued"
                message: "Email submitted for processing. Use the job_id to check status."
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/ingest/image:
    post:
      tags:
        - ingest
      summary: Submit screenshot for order extraction
      description: |
        Submit a screenshot for order extraction. Creates a job that is processed
        asynchronously by the Worker service via Cloud Tasks.

        The image will be analyzed by the input processor agent (with vision capability)
        to extract order information. Duplicate images are detected via SHA-256 hash.
      operationId: ingestImage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestImageRequest"
            examples:
              basic:
                summary: Basic image submission
                value:
                  image_data: "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mNk+M9QDwADhgGAWjR9awAAAABJRU5ErkJggg=="
              withFilename:
                summary: With filename
                value:
                  image_data: "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mNk+M9QDwADhgGAWjR9awAAAABJRU5ErkJggg=="
                  filename: "amazon_order.png"
      responses:
        "200":
          description: Image submitted for processing (or duplicate detected)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
              examples:
                queued:
                  summary: Image queued for processing
                  value:
                    job_id: "550e8400-e29b-41d4-a716-446655440000"
                    source_id: "660e8400-e29b-41d4-a716-446655440001"
                    status: "queued"
                    message: "Image submitted for processing. Use the job_id to check status."
                duplicate:
                  summary: Duplicate image detected
                  value:
                    job_id: ""
                    source_id: "770e8400-e29b-41d4-a716-446655440002"
                    status: "duplicate"
                    message: "Duplicate image detected. Existing order: 880e8400-e29b-41d4-a716-446655440003"
        "400":
          description: Invalid image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: "Invalid base64 image data: Incorrect padding"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        GCP Identity Token for authentication.

        To obtain a token:
        ```bash
        TOKEN=$(gcloud auth print-identity-token)
        ```

        Then include in requests:
        ```
        Authorization: Bearer <token>
        ```

  schemas:
    # OpenAI-compatible Chat Completion schemas
    ChatCompletionRequest:
      type: object
      required:
        - messages
      properties:
        model:
          type: string
          description: Model to use (ignored, uses configured agent model)
          default: "gemini-2.5-flash"
          example: "gemini-2.5-flash"
        messages:
          type: array
          description: List of messages in the conversation
          minItems: 1
          items:
            $ref: "#/components/schemas/ChatMessage"
        stream:
          type: boolean
          description: Whether to stream the response
          default: false
        temperature:
          type: number
          description: Sampling temperature (0-2)
          minimum: 0
          maximum: 2
          nullable: true
        max_tokens:
          type: integer
          description: Maximum tokens to generate
          nullable: true
        user:
          type: string
          description: User identifier for session management
          nullable: true
          example: "user_12345"

    ChatMessage:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum: [system, user, assistant]
          description: The role of the message author
          example: "user"
        content:
          type: string
          description: The content of the message
          example: "Hello! What is Trackable?"

    ChatCompletionResponse:
      type: object
      required:
        - id
        - object
        - created
        - model
        - choices
      properties:
        id:
          type: string
          description: Unique identifier for this completion
          example: "chatcmpl-abc123def456"
        object:
          type: string
          enum: [chat.completion]
          example: "chat.completion"
        created:
          type: integer
          description: Unix timestamp of when the completion was created
          example: 1706300000
        model:
          type: string
          description: Model used for completion
          example: "gemini-2.5-flash"
        choices:
          type: array
          items:
            $ref: "#/components/schemas/ChatCompletionChoice"
        usage:
          $ref: "#/components/schemas/ChatCompletionUsage"

    ChatCompletionChoice:
      type: object
      required:
        - index
        - message
        - finish_reason
      properties:
        index:
          type: integer
          example: 0
        message:
          $ref: "#/components/schemas/ChatCompletionMessage"
        finish_reason:
          type: string
          enum: [stop, length, content_filter]
          nullable: true
          example: "stop"

    ChatCompletionMessage:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum: [assistant]
          example: "assistant"
        content:
          type: string
          example: "Hello! I'm Trackable, your personal shopping assistant."

    ChatCompletionUsage:
      type: object
      properties:
        prompt_tokens:
          type: integer
          example: 10
        completion_tokens:
          type: integer
          example: 15
        total_tokens:
          type: integer
          example: 25

    ChatCompletionChunk:
      type: object
      description: Streaming chunk for chat completion
      required:
        - id
        - object
        - created
        - model
        - choices
      properties:
        id:
          type: string
          example: "chatcmpl-abc123def456"
        object:
          type: string
          enum: [chat.completion.chunk]
          example: "chat.completion.chunk"
        created:
          type: integer
          example: 1706300000
        model:
          type: string
          example: "gemini-2.5-flash"
        choices:
          type: array
          items:
            $ref: "#/components/schemas/ChatCompletionChunkChoice"

    ChatCompletionChunkChoice:
      type: object
      properties:
        index:
          type: integer
          example: 0
        delta:
          $ref: "#/components/schemas/ChatCompletionChunkDelta"
        finish_reason:
          type: string
          enum: [stop, length, content_filter]
          nullable: true

    ChatCompletionChunkDelta:
      type: object
      properties:
        role:
          type: string
          enum: [assistant]
          nullable: true
        content:
          type: string
          nullable: true

    # Ingest API schemas
    IngestEmailRequest:
      type: object
      required:
        - email_content
      properties:
        email_content:
          type: string
          description: Raw email content (EML format or plain text)
          minLength: 1
          example: "From: orders@amazon.com\nSubject: Your order\n\nOrder confirmed!"
        email_subject:
          type: string
          description: Email subject line (optional)
          nullable: true
          example: "Your order has shipped"
        email_from:
          type: string
          description: Sender email address (optional)
          nullable: true
          example: "orders@amazon.com"

    IngestImageRequest:
      type: object
      required:
        - image_data
      properties:
        image_data:
          type: string
          description: Base64 encoded image data
          minLength: 1
        filename:
          type: string
          description: Original filename (optional)
          nullable: true
          example: "screenshot.png"

    IngestResponse:
      type: object
      required:
        - job_id
        - source_id
        - status
        - message
      properties:
        job_id:
          type: string
          format: uuid
          description: Job ID for tracking processing status
          example: "550e8400-e29b-41d4-a716-446655440000"
        source_id:
          type: string
          format: uuid
          description: Source ID for the submitted content
          example: "660e8400-e29b-41d4-a716-446655440001"
        status:
          type: string
          enum: [queued, duplicate]
          description: Initial status (queued or duplicate if image already processed)
          example: "queued"
        message:
          type: string
          description: Human-readable status message
          example: "Email submitted for processing. Use the job_id to check status."

    ClearSessionResponse:
      type: object
      required:
        - message
        - user
      properties:
        message:
          type: string
          description: Status message indicating whether session was cleared
          example: "Session cleared"
        user:
          type: string
          description: User identifier
          example: "user_12345"

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message describing what went wrong
          example: "Chat completion failed: Internal error"

  examples:
    BasicChatRequest:
      summary: Basic chat completion request
      value:
        model: "gemini-2.5-flash"
        messages:
          - role: "user"
            content: "Hello! What is Trackable?"
        user: "user_12345"

    StreamingChatRequest:
      summary: Streaming chat completion request
      value:
        model: "gemini-2.5-flash"
        messages:
          - role: "user"
            content: "Tell me about order tracking"
        stream: true
        user: "user_12345"

    ChatCompletionExample:
      summary: Successful chat completion response
      value:
        id: "chatcmpl-abc123def456"
        object: "chat.completion"
        created: 1706300000
        model: "gemini-2.5-flash"
        choices:
          - index: 0
            message:
              role: "assistant"
              content: "Hello! I'm Trackable, your personal shopping assistant."
            finish_reason: "stop"
        usage:
          prompt_tokens: 10
          completion_tokens: 15
          total_tokens: 25